Merge transitive and direct reverse dependencies, close sbt/stb#5484
Motivation:

Direct reverse dependencies Map is added into the transitive reverse dependencies Map, so any dependency existing in both Maps loses transitive callers.

This behavior is not consistent with the ivy based lm and is not desirable: the goal is to have the complete callers tree.

Modification:

Merge the 2 maps instead of adding and lose values for conflicting keys.

Result:

Transitive callers are not lost when dependency is also pulled directly.