testing SbtCoursierCache leak
however if I revert the fix the test is still green. I added
a few printlns and can see that the SbtCoursierCache
instance after the clean is different.

Does this happen because of the reload in a scripted test is
different? or maybe the problem I reported only manifest in
more complex scenarios?